perf(brain_client): run the WebSocket transport in-process
Fuse the ws_client transport into brain_client_node so cloud messages — the
high-rate image/vision payloads on the perception path — no longer cross the
ws_messages/ws_outgoing String topics or get re-serialized in a second process.

- comms/ws_manager.py: WebSocketManager — the connection orchestration (config
  validation, reconnect thread, status publishing, backend-config swaps,
  READY_FOR_CONNECTION routing) extracted from WSClientNode, node-agnostic.
- comms/websocket.py: WSBridge gains an in-process transport mode — send_message
  hands the message object straight to the manager (one serialization, no topic);
  dispatch() serves as the manager's on_incoming.
- nodes/brain_client_node.py: owns a WebSocketManager in-process by default
  (inprocess_websocket param, default true); falls back to the topic bridge when
  false.
- nodes/ws_client.py: thin wrapper over WebSocketManager (topic mode) for the
  standalone / fallback case.
- Launch files no longer spawn a separate ws_client node.
- test_fake_cloud_loop runs fused (no ws_client); test_pose_image pins
  inprocess_websocket:=false since it injects/observes via ws_messages/ws_outgoing.
- Also add the missing rclpy.ok() shutdown guard to input_manager for parity.

Verified in innate-os-test:fullbuild: colcon build, 36 pure unit tests, all 4
node executables construct, test_fake_cloud_loop 5/5 (fused, full directive->
primitive->completion loop over the in-process transport), test_pose_image 3/3.
