+++ ros2_ws/src/brain/brain_client/launch/brain_client.launch.py
@@ -184,19 +184,8 @@ def generate_launch_description():
             openai_realtime_url_arg,
             openai_transcribe_model_arg,
             brain_client_node,
-            # Launch the WSClientNode (handles actual WebSocket connection)
-            Node(
-                package="brain_client",
-                executable="ws_client.py",
-                name="ws_client_node",
-                output="screen",
-                parameters=[
-                    {
-                        "websocket_uri": LaunchConfiguration("websocket_uri"),
-                        "token": LaunchConfiguration("token"),
-                    }
-                ],
-            ),
+            # The WebSocket connection runs in-process inside brain_client_node
+            # (inprocess_websocket defaults to true); no separate ws_client node.
             Node(
                 package="brain_client",
                 executable="skills_server.py",
